Add two ways of showing the fingerprint Printing the fingerprint on startup to logging.INFO has two drawbacks:
1. With no default handlers, the user will not see INFO-level logging messages, which can be confusing.
2. If you don't want the fingerprint in your logs each time, there was no option to turn it off without raising the logging level.

This change will print the fingerprint to stdout by default, and this option can be turned off if so desired.
A user can also get their fingerprint using a command-line utility that they can run before running or building their client for the first time.

This is documented under Message Signing > Certificate Fingerprints.