Made 'measurement' objects more consistent across different parts of OpenLEADR
The member names of the 'measurement' objects or dicts have been changed to be consistent everywhere:
- item_name -> name
- item_description -> description
- item_units -> unit
- si_scale_code -> scale

This way, the parameters to client.add_report() are consistent with the Measurement object and the dicts that are passed around.

Additionally, there is extra validation to prevent sending invalid measurements, and hints to correct any mistakes.

Signed-off-by: Stan Janssen <stan.janssen@elaad.nl>