Improved flexibility for assigning targets to an Event.
This adds the possibility to add targets as a dict of targets_by_type. This makes the instantiation of the Event more in line with the representation of the Event.

If you don't supply any targets, the ven_id becomes the target.

It also adds support for setting the Market Context of the Event.

Signed-off-by: Stan Janssen <stan.janssen@elaad.nl>