Bug: create a single report: add breaking test
When you create a single report on the VEN client,
the VEN client correctly translates that to an XML message,
but the resulting XML is incorrectly deserialized into an object:
the value of 'pending_reports' becomes a bare dict instead of a list with one
dict.

This commit adds the breaking test; next commit adds the fix.