Fixed report_request_id parsing bug
This is an alternative to the fix offered in !116, which solves a
problem where if there is only one pending report (with a single
report_request_id), it would not get turned into a proper list.

The solution offered in !116, which uses a more expressive style to
reason about the shape of the contents looks promising, and worth
exploring as part of #17 as well.

For now, this commit solves the reported problem without additional
dependencies.